tests/test_secure_auth/test_requirements.txt
# Test requirements for secure_auth PostgreSQL tests
pytest>=7.0.0
pytest-postgresql>=4.0.0
pytest-asyncio>=0.21.0
psycopg2-binary>=2.9.0